[Engine] Consistently handle Display IDs above INT64_MAX
Dart ints are signed 64-bit, so a display ID above INT64_MAX can't be
represented in Dart. What actually happened with such an ID depended on
the toolchain.

tonic's DartConverter<unsigned long long> asserted the range, while
every other integer type has no check. So it came down to what size_t
is: on MSVC x64 debug builds aborted, on LP64 the same ID silently
wrapped negative. This removes the specialization, as suggested by
knopp, so all platforms behave the same now: values above INT64_MAX
become the corresponding negative Dart integer, distinct values remain
distinct. The dropped constants have no other users and nothing relied
on the assert.

DisplayId itself was size_t, but the ABI says uint64_t, so 32-bit
builds truncated IDs at the embedder boundary. It is int64_t now (like
FlutterViewId), with one explicit cast at the boundary, and
ViewportMetrics::display_id got the same change. The tonic change
alone fixes the crash, the typedef fixes the 32-bit narrowing.

This replaces the HMONITOR masking this PR originally proposed.

Test:
EmbedderTest.SetSingleDisplayConfigurationWithDisplayIdAboveInt64Max
sends 0xFFFFFFFFE02E16A5, a real sign-extended HMONITOR, and checks it
reaches Dart as the negative equivalent through both the display list
and the window metrics.

Fixes #189564

+++ engine/src/flutter/lib/ui/window/viewport_metrics.h
@@ -5,6 +5,7 @@
 #ifndef FLUTTER_LIB_UI_WINDOW_VIEWPORT_METRICS_H_
 #define FLUTTER_LIB_UI_WINDOW_VIEWPORT_METRICS_H_
 
+#include <cstdint>
 #include <ostream>
 #include <vector>
 
@@ -16,7 +17,7 @@ struct ViewportMetrics {
                   double p_physical_width,
                   double p_physical_height,
                   double p_physical_touch_slop,
-                  size_t display_id);
+                  int64_t display_id);
   ViewportMetrics(double p_device_pixel_ratio,
                   double p_physical_width,
                   double p_physical_height,
@@ -40,7 +41,7 @@ struct ViewportMetrics {
                   const std::vector<double>& p_physical_display_features_bounds,
                   const std::vector<int>& p_physical_display_features_type,
                   const std::vector<int>& p_physical_display_features_state,
-                  size_t p_display_id,
+                  int64_t p_display_id,
                   double p_physical_display_corner_radius_top_left,
                   double p_physical_display_corner_radius_top_right,
                   double p_physical_display_corner_radius_bottom_right,
@@ -68,7 +69,7 @@ struct ViewportMetrics {
   std::vector<double> physical_display_features_bounds;
   std::vector<int> physical_display_features_type;
   std::vector<int> physical_display_features_state;
-  size_t display_id = 0;
+  int64_t display_id = 0;
   double physical_display_corner_radius_top_left = -1.0;
   double physical_display_corner_radius_top_right = -1.0;
   double physical_display_corner_radius_bottom_right = -1.0;
